refactor(configs): remove `.babel` from file names that don't need it
refactor(configs): remove `.babel` from file names that don't need it
